rename: btnText 네이밍 및 파일 위치 변경
