+++ src/shared/components/v2/navBar/LogoNavBar.tsx
@@ -3,7 +3,7 @@ import logotypeBlack from '@assets/v2/images/LogotypeBlack.svg';
 import logotypeWhite from '@assets/v2/images/LogotypeWhite.svg';
 
 import * as styles from './LogoNavBar.css';
-import BtnText from '../btnText/BtnText';
+import TextButton from '../btnText/TextButton';
 
 type AuthSlot = 'none' | 'login' | 'profile';
 type Page = 'landing' | 'home';
@@ -47,9 +47,9 @@ const LogoNavBar = ({
         )}
         {authSlot === 'login' && (
           <div className={styles.actionContainer}>
-            <BtnText color="primary" size="s" onClick={onLoginClick}>
+            <TextButton color="primary" size="s" onClick={onLoginClick}>
               로그인
-            </BtnText>
+            </TextButton>
           </div>
         )}
         {authSlot === 'profile' && (
